Moves kubectl get subcommand to staging
