+++ pkg/kubectl/cmd/BUILD
@@ -12,7 +12,6 @@ go_library(
         "//pkg/kubectl/cmd/auth:go_default_library",
         "//pkg/kubectl/cmd/convert:go_default_library",
         "//pkg/kubectl/cmd/cp:go_default_library",
-        "//pkg/kubectl/cmd/get:go_default_library",
         "//staging/src/k8s.io/cli-runtime/pkg/genericclioptions:go_default_library",
         "//staging/src/k8s.io/client-go/tools/clientcmd:go_default_library",
         "//staging/src/k8s.io/component-base/cli/flag:go_default_library",
@@ -35,6 +34,7 @@ go_library(
         "//staging/src/k8s.io/kubectl/pkg/cmd/exec:go_default_library",
         "//staging/src/k8s.io/kubectl/pkg/cmd/explain:go_default_library",
         "//staging/src/k8s.io/kubectl/pkg/cmd/expose:go_default_library",
+        "//staging/src/k8s.io/kubectl/pkg/cmd/get:go_default_library",
         "//staging/src/k8s.io/kubectl/pkg/cmd/kustomize:go_default_library",
         "//staging/src/k8s.io/kubectl/pkg/cmd/label:go_default_library",
         "//staging/src/k8s.io/kubectl/pkg/cmd/logs:go_default_library",
@@ -86,7 +86,6 @@ filegroup(
         "//pkg/kubectl/cmd/auth:all-srcs",
         "//pkg/kubectl/cmd/convert:all-srcs",
         "//pkg/kubectl/cmd/cp:all-srcs",
-        "//pkg/kubectl/cmd/get:all-srcs",
         "//pkg/kubectl/cmd/plugin/testdata:all-srcs",
     ],
     tags = ["automanaged"],
